+++ src/main/java/chungbazi/chungbazi_be/domain/auth/service/KakaoAuthService.java
@@ -9,7 +9,6 @@ import chungbazi.chungbazi_be.domain.auth.jwt.SecurityUtils;
 import chungbazi.chungbazi_be.domain.auth.jwt.TokenGenerator;
 import chungbazi.chungbazi_be.domain.notification.service.FCMTokenService;
 import chungbazi.chungbazi_be.domain.user.entity.User;
-import chungbazi.chungbazi_be.domain.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -17,9 +16,9 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class KakaoAuthService {
 
-    private final UserService userService;
+    private final UserAuthService userAuthService;
     private final TokenGenerator tokenGenerator;
-    private final AuthTokenService authTokenService;
+    private final TokenAuthService tokenAuthService;
     private final FCMTokenService fcmTokenService;
     private final KakaoAuthConverter kakaoAuthConverter;
     private final JwtProvider jwtProvider;
@@ -27,10 +26,10 @@ public class KakaoAuthService {
 
     // 로그인 및 회원가입 처리
     public TokenDTO loginUser(TokenRequestDTO.LoginTokenRequestDTO request) {
-        User user = userService.findOrCreateMember(request);
-        boolean isFirst = userService.determineIsFirst(user);
+        User user = userAuthService.findOrCreateMember(request);
+        boolean isFirst = userAuthService.determineIsFirst(user);
         TokenDTO tokenDTO = tokenGenerator.generate(user.getId(), user.getName(), isFirst);
-        authTokenService.saveRefreshToken(user.getId(), tokenDTO.getRefreshToken(), tokenDTO.getRefreshExp());
+        tokenAuthService.saveRefreshToken(user.getId(), tokenDTO.getRefreshToken(), tokenDTO.getRefreshExp());
         fcmTokenService.saveFcmToken(user.getId(), request.getFcmToken());
         return tokenDTO;
     }
@@ -38,28 +37,28 @@ public class KakaoAuthService {
     // JWT 토큰 관련 처리
     public TokenDTO recreateAccessToken(String refreshToken) {
         Long userId = jwtProvider.getUserIdFromToken(refreshToken);
-        authTokenService.validateRefreshToken(userId, refreshToken);
+        tokenAuthService.validateRefreshToken(userId, refreshToken);
 
-        User user = userService.getUserById(userId);
-        authTokenService.addToBlackList(refreshToken,"expired", 3600L);
+        User user = userAuthService.getUserById(userId);
+        tokenAuthService.addToBlackList(refreshToken,"expired", 3600L);
         return tokenGenerator.generate(userId, user.getName(), false);
     }
 
     // 로그아웃
     public void logoutUser(String token) {
-        authTokenService.validateNotBlackListed(token);
+        tokenAuthService.validateNotBlackListed(token);
         Long userId = SecurityUtils.getUserId();
-        authTokenService.addToBlackList(token, "logout", 3600L);
-        authTokenService.deleteRefreshToken(userId);
+        tokenAuthService.addToBlackList(token, "logout", 3600L);
+        tokenAuthService.deleteRefreshToken(userId);
     }
 
     // 회원 탈퇴
     public void deleteUserAccount(String token) {
-        authTokenService.validateNotBlackListed(token);
+        tokenAuthService.validateNotBlackListed(token);
         Long userId = SecurityUtils.getUserId();
-        authTokenService.addToBlackList(token, "delete-account", 3600L);
-        authTokenService.deleteRefreshToken(userId);
-        userService.deleteUser(userId);
+        tokenAuthService.addToBlackList(token, "delete-account", 3600L);
+        tokenAuthService.deleteRefreshToken(userId);
+        userAuthService.deleteUser(userId);
     }
 
     // 응답
