+++ src/main/java/chungbazi/chungbazi_be/domain/user/service/UserService.java
@@ -1,7 +1,6 @@
 package chungbazi.chungbazi_be.domain.user.service;
 
 
-import chungbazi.chungbazi_be.domain.auth.dto.TokenRequestDTO;
 import chungbazi.chungbazi_be.domain.auth.jwt.SecurityUtils;
 import chungbazi.chungbazi_be.domain.user.converter.UserConverter;
 import chungbazi.chungbazi_be.domain.user.dto.UserRequestDTO;
@@ -19,9 +18,9 @@ import chungbazi.chungbazi_be.global.s3.S3Manager;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
-import org.springframework.web.multipart.MultipartFile;
 import java.util.Optional;
 
 
@@ -133,37 +132,4 @@ public class UserService {
         }
     }
 
-    public User findOrCreateMember(TokenRequestDTO.LoginTokenRequestDTO request) {
-        return userRepository.findByEmail(request.getEmail())
-                .map(existingUser -> {
-                    if (existingUser.isDeleted()) {
-                        throw new BadRequestHandler(ErrorStatus.DEACTIVATED_ACCOUNT);
-                    }
-                    return existingUser;
-                })
-                .orElseGet(() -> createNewUser(request));
-    }
-    public User createNewUser(TokenRequestDTO.LoginTokenRequestDTO request) {
-        User user = User.builder()
-                .email(request.getEmail())
-                .name(request.getName())
-                .build();
-        return userRepository.save(user);
-    }
-
-    public boolean determineIsFirst(User user) {
-        return !user.isSurveyStatus();
-    }
-
-    public User getUserById(Long userId) {
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundHandler(ErrorStatus.NOT_FOUND_USER));
-    }
-
-    public void deleteUser(Long userId) {
-        userRepository.findById(userId).ifPresent(user -> {
-            user.updateIsDeleted(true);
-            userRepository.save(user);
-        });
-    }
 }
